feat: improve step nodes design and extract base styles in a separate component
